designorganapp/requirements/lista_de_paquetes.txt
asgiref==3.3.1
Django==3.1.4
django-crum==0.7.9
django-widget-tweaks==1.4.8
freeze==3.0
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1








html5lib==1.0.1
mysqlclient==1.4.6

PyPDF2==1.26.0

reportlab==3.5.42


webencodings==0.5.1
xhtml2pdf==0.2.4